Removed $scope.save method
The $scope.save was planned to send an invoice via post to a server to
save it there but isn't currently used so I removed it.  I was 1/2
working on a node.js server component and it looks like this part of
that code got into the branch that wasn't for the server side stuff.